+++ internals/npm/js.prod.js
@@ -9,7 +9,7 @@
 process.env.BABEL_ENV = "production";
 process.env.NODE_ENV = "production";
 
-const fs = require("fs-extra"); // https://www.npmjs.com/package/fs-extra
+//const fs = require("fs-extra"); // https://www.npmjs.com/package/fs-extra
 const chalk = require("chalk"); // https://www.npmjs.com/package/chalk
 const webpack = require("webpack"); // https://www.npmjs.com/package/webpack
 const webpackConfig = require("../webpack/config.prod");