+++ internals/npm/seed.js
@@ -5,10 +5,12 @@
  * @type {Node.js}
  *
  **********************************************************************************************************************/
-const mongoose = require("mongoose");
+const mongoose = require("mongoose"); // http://mongoosejs.com/
 const path = require("path"); // https://nodejs.org/api/path.html
+const chalk = require("chalk"); // https://www.npmjs.com/package/chalk
 const fs = require("fs");
 const config = require("../config");
+const seed = require("../../server/handlers/game/seed");
 const logger = require("../utils/clientlogger");
 require("dotenv").config({
     path: path.join(config.projectRoot, "server", "config", ".env")
@@ -30,24 +32,10 @@ const stars = require("../mockdata/stars");
 const planets = require("../mockdata/planets");
 const turns = require("../mockdata/turns");
 
-// add maxPlayers to game == number of player systems
-games.forEach( game => {
-    let gameStars = stars.filter(star => star.homeSystem && star.game === game._id);
-    game.maxPlayers = gameStars.length;
-    logger.info(`game ${game.number} has ${gameStars.length} maxplayers`);
-});
-
-
-// http://mongoosejs.com/docs/connections.html#use-mongo-client
-mongoose.connect(process.env.DATABASE, {
-    useMongoClient: true
-});
-mongoose.Promise = global.Promise;
-
 /*
  * prune database and throw everything away.
  */
-async function pruneDatabase() {
+const pruneDatabase = async () => {
     logger.info("[node] deleting collections ...");
     try {
         await Game.remove();
@@ -70,12 +58,12 @@ async function pruneDatabase() {
         logger.error(e);
     }
     process.exit(0);
-}
+};
 
 /*
  * seed database with mock json data
  */
-async function seedDatabase() {
+const seedDatabase = async () => {
     logger.info("[node] start seeding database");
     try {
         logger.debug(`[node] inserting ${stars.length} stars.`);
@@ -90,13 +78,74 @@ async function seedDatabase() {
         await Planet.insertMany(planets);
         logger.debug(`[node] inserting ${turns.length} turns.`);
         await Turn.insertMany(turns);
-        logger.success("[node] done, database seeded!");
+        logger.debug("[node] done inserting.");
+        logger.success("[node] finished seeding database.");
     } catch (e) {
         logger.error("[node] error while seeding database.");
         logger.error(e);
     }
     process.exit(0);
-}
+};
+
+/*
+ * since maxplayers are determined by available homeSystems (which depends on size and distance)
+ * we need to assign them now that stars have been created
+ */
+logger.info(
+    `[node] modifying maxplayers for ${chalk.red(games.length)} Games.`
+);
+games.forEach(game => {
+    let gameStars = stars.filter(
+        star => star.homeSystem && star.game === game._id
+    );
+    game.maxPlayers = gameStars.length;
+    logger.info(
+        `game ${chalk.red("g" + game.number)} set to ${chalk.yellow(
+            gameStars.length
+        )} maxplayers`
+    );
+});
+
+/*
+ * assign homeSystems to players. this is normally done during enlisting.
+ */
+logger.info(
+    `[node] seeding homesystems for ${chalk.red(players.length)} Players.`
+);
+players.forEach(player => {
+    const gameStars = stars.filter(
+        star =>
+            star.game === player.game &&
+            star.homeSystem === true &&
+            star.owner === undefined
+    );
+    const homeSystem = seed.assignRandomStar(gameStars);
+    let index = 0;
+    stars.forEach(star => {
+        if (
+            player.game === star.game &&
+            star.name === homeSystem.name &&
+            star.coordX === homeSystem.coordX &&
+            star.coordY === homeSystem.coordY
+        ) {
+            stars[index].owner = player._id;
+        }
+        index++;
+    });
+    logger.info(
+        `[node] chosen Star ${chalk.yellow(
+            homeSystem.name
+        )} for player ${chalk.red("[" + player.ticker + "]")} ${chalk.cyan(
+            player.name
+        )}`
+    );
+});
+
+// http://mongoosejs.com/docs/connections.html#use-mongo-client
+mongoose.connect(process.env.DATABASE, {
+    useMongoClient: true
+});
+mongoose.Promise = global.Promise;
 
 if (process.argv.includes("--prune")) {
     try {